% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{import_items}
\alias{import_items}
\title{Import \emph{only} the collection options from a  json config file}
\usage{
import_items(file)
}
\description{
Import \emph{only} the collection options from a  json config file
}
