% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{project_options}
\alias{project_options}
\title{Project options S3 object}
\usage{
project_options(
  raw = "./data/raw",
  interim = "./data/interim",
  processed = "./outputs/processed",
  setup = FALSE
)
}
\description{
Project options S3 object
}
