% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_read.R
\name{raster_make_list}
\alias{raster_make_list}
\title{Create a list of raster files in the format required}
\usage{
raster_make_list(files)
}
\arguments{
\item{files}{named character vector; Raster file paths with names
corresponding to the type of raster ("categorical" or "continuous"}
}
\value{
A list with two elements: 'files' containing the file paths and
'types' containing the names of the files.
}
\description{
Create a list of raster files in the format required
}
\examples{
files <- c("categorical" = "path/to/file1.tif")
rasters <- raster_make_list(files)
}
