% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_read.R
\name{raster_read}
\alias{raster_read}
\title{Read a list of raster files}
\usage{
raster_read(raster_list)
}
\arguments{
\item{raster_list}{list;  list of raster files created by
\code{\link{raster_make_list}}}
}
\value{
A list of spatRaster objects
}
\description{
Read a list of raster files
}
